Data layer control now is part of map layer control, also custom radio buttons are fixed
