+++ client/scripts/stores/arcgisLayerStore.js
@@ -63,7 +63,7 @@ module.exports = Reflux.createStore({
 		var theLayer = _.findWhere(this.state.layers, {'id': id});
 		var isFeature=theLayer.type=='Feature Service';
 
-		if (property == 'moveDown' && isTile) {
+		if (property == 'moveDown' && !isFeature) {
 			var currentZindex = theLayer.zIndex;
 			if (currentZindex > 0) {
 				var newZindex = currentZindex - 1;
