+++ client/styles/main.scss
@@ -19,3 +19,4 @@
 @import 'modules/login';
 @import 'modules/language-nav';
 
+@import 'modules/draft'; //dev styles please replace
