+++ client/styles/modules/draft.scss
@@ -1,1 +1,25 @@
 /*add here all styles used during development later they will be replaced by the final ones */
+
+
+	.selectable-radio{
+		  border-radius: 50% !important;
+		  background-clip:none;
+
+		font-size: 1.5em !important; 
+		width: 5px;
+		height: 5px;
+		border: 1px solid #cccccc;
+		background: #ffffff;
+		float: left;
+		margin-right: 0.2em;
+		margin-top: 1px;
+		width:13px;
+		height:13px;
+		
+		.fa{
+			font-size:0.7em;
+			margin-left: 2px;
+			margin-top: -2px;
+			float: left;
+		}
+	}
